pyproject.toml
[tool.poetry]
name = "premier-league-predictor"
version = "0.1.0"
description = "Predict Premier League match outcomes using RF model"
authors = ["User <user@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.2.2"
numpy = "^1.26.4"
scikit-learn = "^1.4.2"
requests = "^2.31.0"
joblib = "^1.4.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
